requirements.txt
# InsightVault Requirements

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.26.0

# Database
psycopg[binary,pool]>=3.1.16

# Vector Database
pgvector

# AI & Embeddings
openai>=1.10.0

# Multi-modal Processing
PyMuPDF>=1.23.0              # PDF parsing
python-docx>=1.1.0           # Word document parsing
python-pptx>=0.6.23          # PowerPoint parsing
Pillow>=10.2.0               # Image processing
pytesseract>=0.3.10          # OCR (local)
opencv-python>=4.9.0         # Image preprocessing
openai-whisper>=20231117     # Audio transcription (optional)

# Data Processing
beautifulsoup4>=4.12.0       # HTML parsing
feedparser>=6.0.10           # RSS feed parsing
markdownify>=0.11.0          # HTML to Markdown conversion

# Utilities
Jinja2>=3.1.3
pydantic>=2.5.0
